app: integer-ify position/offset members of GimpSourceCore
We don't support subpixel source sampling, so there's no use in
pretending that we do.  Demoting everything to int as soon as
possible helps guarantee that these values are at least rounded
properly and in fewer places.

Make sure we always round coordinates down, and not toward zero.

Keep using floats only in the signatures of the relevant PDB
functions.

+++ app/paint/gimpsourcecore.h
@@ -39,14 +39,14 @@ struct _GimpSourceCore
   gboolean       set_source;
 
   GimpDrawable  *src_drawable;
-  gdouble        src_x;
-  gdouble        src_y;
+  gint           src_x;
+  gint           src_y;
 
-  gdouble        orig_src_x;
-  gdouble        orig_src_y;
+  gint           orig_src_x;
+  gint           orig_src_y;
 
-  gdouble        offset_x;
-  gdouble        offset_y;
+  gint           offset_x;
+  gint           offset_y;
   gboolean       first_stroke;
 };
